feat: generate article summary using gemini

+++ Clients/Models/GenerationConfig.cs
@@ -1,7 +1,7 @@
 namespace bsky.bot.Clients.Models;
 
 public readonly record struct GenerationConfig(
-    int temperature,
+    double temperature,
     double topK,
     double topP,
     int maxOutputTokens,
